PLAT-1069 | move signups open validation to middleware
