--- app/Exceptions/SignupForProductIsClosedException.php
@@ -1,7 +1,0 @@
-<?php
-
-namespace App\Exceptions;
-
-use Illuminate\Http\Exceptions\HttpResponseException;
-
-class SignupForProductIsClosedException extends HttpResponseException {}
